+++ packages/accounts/app/components/AccountItem.vue
@@ -1,0 +1,8 @@
+<script setup lang="ts">
+</script>
+
+<template>
+  <div class="bg-green-600">
+    New Item Accounts Layer
+  </div>
+</template>