+++ packages/accounts/app/components/AccountItem2.vue
@@ -1,0 +1,8 @@
+<script setup lang="ts">
+</script>
+
+<template>
+  <div class="bg-red-800/50">
+    New Item Accounts Layer
+  </div>
+</template>