+++ packages/base-ui/.playground/nuxt.config.ts
@@ -1,4 +1,7 @@
 export default defineNuxtConfig({
   extends: ['..'],
-  modules: []
+  modules: [],
+  css: [
+    '~/assets/css/tw-base-playground.css'
+  ]
 })