+++ packages/base-ui/app/components/BaseItem.vue
@@ -1,0 +1,8 @@
+<script setup lang="ts">
+</script>
+
+<template>
+  <div class="bg-midnightBlue-950">
+    Base Layer Component
+  </div>
+</template>