+++ packages/base-ui/app/components/NewItem.vue
@@ -1,0 +1,15 @@
+<script setup lang="ts">
+const something = ref(false)
+</script>
+
+<template>
+  <div class="bg-yellow-500">
+    I am a new item
+    <br>
+    {{ something }}
+    <br>
+    more stuff here
+    <br>
+    again more stuff
+  </div>
+</template>