+++ packages/pay/CHANGELOG.md
@@ -1,5 +1,70 @@
 # @dwol/pay-layer
 
+## 0.0.17-rc.8
+
+### Patch Changes
+
+- [#104](https://github.com/deetz99/nuxt-layers-monorepo/pull/104) [`d6d119d`](https://github.com/deetz99/nuxt-layers-monorepo/commit/d6d119d640c5e9e732d8ab0e4e90aa3ef555d63f) Thanks [@deetz99](https://github.com/deetz99)! - Chore(release): enter rc version
+
+- Updated dependencies [[`d6d119d`](https://github.com/deetz99/nuxt-layers-monorepo/commit/d6d119d640c5e9e732d8ab0e4e90aa3ef555d63f)]:
+  - @dwol/accounts-layer@0.0.19-rc.8
+
+## 0.0.17-beta.7
+
+### Patch Changes
+
+- Updated dependencies [[`85eec27`](https://github.com/deetz99/nuxt-layers-monorepo/commit/85eec278cf28cd3e5f7cefdebd26c5364e146121)]:
+  - @dwol/accounts-layer@0.0.19-beta.7
+
+## 0.0.17-beta.6
+
+### Patch Changes
+
+- Updated dependencies [[`0f317cc`](https://github.com/deetz99/nuxt-layers-monorepo/commit/0f317ccf9e43651b25bc99988722e41551803ff7)]:
+  - @dwol/accounts-layer@0.0.19-beta.6
+
+## 0.0.17-beta.5
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @dwol/accounts-layer@0.0.19-beta.5
+
+## 0.0.17-beta.4
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @dwol/accounts-layer@0.0.19-beta.4
+
+## 0.0.17-beta.3
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @dwol/accounts-layer@0.0.19-beta.3
+
+## 0.0.17-beta.2
+
+### Patch Changes
+
+- Updated dependencies [[`db173be`](https://github.com/deetz99/nuxt-layers-monorepo/commit/db173bef96483c863c40460a0df10a3d21a56d9c)]:
+  - @dwol/accounts-layer@0.0.19-beta.2
+
+## 0.0.17-beta.1
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @dwol/accounts-layer@0.0.19-beta.1
+
+## 0.0.17-beta.0
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @dwol/accounts-layer@0.0.19-beta.0
+
 ## 0.0.16
 
 ### Patch Changes